Adding compatibility for Django 1.9
